requirements.txt
# Listening for audio input.
pvrecorder

# Listening for wake word in the audio input.
pvporcupine

# Listening for speech activity/silence in the audio input.
pvcobra

# Converting incoming user input from speech-to-text.
numpy
pywhispercpp
pvcheetah

# Sending user input text to a model
ollama

# Converting outgoing model response from text-to-speech.
pvorca
# kokoro-tts
pvspeaker

# Prompt management.
pyyaml

# Tooling.
langchain-core
langchain-ollama
